Conflicts:
	gwt-cordova/pom.xml
	gwt-cordova/setup-cordova.sh
	gwt-cordova/src/main/java/nl/ru/languageininteraction/language/client/view/SimpleView.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/AppController.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/presenter/FeedbackPresenter.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/presenter/MetadataPresenter.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/presenter/MoreInfoPresenter.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/presenter/RegisterPresenter.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/presenter/ReportPresenter.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/view/RegisterView.java
	gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/view/ReportView.java
	gwt-cordova/src/main/resources/nl/ru/languageininteraction/language/client/Messages.properties

+++ gwt-cordova/src/main/java/nl/ru/languageininteraction/synaesthesia/client/model/GroupScoreData.java
@@ -26,12 +26,12 @@ import java.util.List;
 public class GroupScoreData {
 
     private final List<ScoreData> scoreDataList;
-    private final double score;
+    private final float score;
     private final double accuracy;
     private final double meanReactionTime;
     private final double reactionTimeDeviation;
 
-    public GroupScoreData(List<ScoreData> scoreDataList, double score, double accuracy, double meanReactionTime, double ReactionTimeDeviation) {
+    public GroupScoreData(List<ScoreData> scoreDataList, float score, double accuracy, double meanReactionTime, double ReactionTimeDeviation) {
         this.scoreDataList = scoreDataList;
         this.score = score;
         this.accuracy = accuracy;
